% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5ad2rds.R
\name{h5ad2rds}
\alias{h5ad2rds}
\title{Convert H5AD file to RDS format}
\usage{
h5ad2rds(h5ad_path)
}
\arguments{
\item{h5ad_path}{The path to the H5AD file.}
}
\value{
String, the path to the generated RDS file.
}
\description{
This function converts a H5AD file to RDS format.
}
\details{
This function creates Seurat5 object via h5ad_to_seurat and saves it to RDS file named
the same way as input file but with .rds extension.
}
\examples{
h5ad2rds("/path/to/h5ad_file.h5ad")

}
