% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h5ad2rds.R
\name{h5ad_to_seurat}
\alias{h5ad_to_seurat}
\title{Convert H5AD file to Seurat object}
\usage{
h5ad_to_seurat(h5ad_path)
}
\arguments{
\item{h5ad_path}{The file path to the H5AD file.}
}
\value{
A Seurat5 object.
}
\description{
This function creates a Seurat object from AnnData (h5ad) file.
}
\details{
This function reads the h5ad file using 'rhdf5' package
and create named list with all fields existing in the file. After that,
Seurat5 object is created from the list.
The package was created for Cell Annotation Platform (CAP) project and expects that AnnData file was created with
python AnnData package of version 0.8+

Resulting Seurat object will have the only assay named 'RNA'.
}
\examples{
seurat_obj <- h5ad_to_seurat("/path/to/h5ad_file.h5ad")
str(seurat_obj)
}
